refactor: simplify state service traits using ZebraService supertrait
