+++ zebra-node-services/Cargo.toml
@@ -31,6 +31,7 @@ rpc-client = [
 
 [dependencies]
 zebra-chain = { path = "../zebra-chain" , version = "2.0.0" }
+tower = { workspace = true }
 
 # Optional dependencies
