+++ zebra-state/Cargo.toml
@@ -77,6 +77,7 @@ sapling-crypto = { workspace = true }
 elasticsearch = { workspace = true, features = ["rustls-tls"], optional = true }
 serde_json = { workspace = true, optional = true }
 
+zebra-node-services = { path = "../zebra-node-services", version = "1.0.1" }
 zebra-chain = { path = "../zebra-chain", version = "2.0.0", features = ["async-error"] }
 
 # prod feature progress-bar
